Test fixes after laziness in cred/authkey evaluation ended
The lambda wrapper represents a change already done earlier in
"remote_cred_cb takes a Callable that returns the remote credentials"
(just never hit the test cases).

The cred_id[ri] setting is not so much done to have that available (it's
not even stored currently), but to ensure that this step (that usually
happens when parsing an incoming message) takes its side effects (of
evaluating the credentials callback and storing remote cred and authkey)
before these are used in the test evaluations.